src/requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
pyyaml==6.0.3
